requirements.txt
aiohttp==3.10.10
einops==0.8.0
GitPython==3.1.43
GitPython==3.1.43
imageio_ffmpeg==0.5.1
kornia==0.7.4
numexpr==2.10.1
numpy==2.1.3
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
psutil==6.1.0
pystray==0.19.5
pytest==8.3.3
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
spandrel==0.4.0
torch==2.5.0+cu124
torchaudio==2.5.0+cu124
torchsde==0.2.6
torchvision==0.20.0+cu124
tqdm==4.66.5
transformers==4.46.0
typer==0.13.0
typing_extensions==4.12.2
websocket_client==1.8.0
open-webui==0.3.35
